test: decoding을 확인할 수 있는 test case 수정
- 기존의 test 에서 URL을 찾지 못하여 tesetBundle 활용하여 변경해줌
